End button & more naming convention
Added an end game button to the main menu and renamed the opponent and
scenario int's in Global to follow proper naming convention (by the way
guys, to properly rename a variable, right click where you declared it,
choose refactor, the rename. It'll change the name anywhere in the
program that you used that variable this way, so you don't have to
manually do it. Also delete variables this way, it has a safely delete
option).