Design Elements Revised and Updated
Changed elements of the main menu and map screen. Added a map to map
screen. Added a menu to map screen. Rearranged the main menu. Added new
items to Help screen. Created a "design elements" folder outside the
project and an "images" folder inside the project package, the former
for storing any external objects required while the latter is for the
external images used within the game.