Team Tweaks
Changed variable Global.chosenCharacter to Global.chosenTeam, made other
modifications in regards to the variable names for teams (basically just
changing everywhere that referenced the old name to reference the new
name, since refractoring seemed to not work for some reason). Also
modified the aesthetics of the team selection screen.