Background and Menus
Finally added the goddamn background image for the main menu and figured
out a very easy way of adding backgrounds to the rest of the forms if we
so choose to.

Created a popup menu that will appear beside whichever map node is
clicked, that will (hopefully) be filled with dynamic content.