New Combat Unit Detection System
Changed the combat unit detection system, now using a rectangular array.
This way the map can be given the id, location, size, and any other
variables needed, all with only having to call to the scenario class
only once (although, for now, it still has to make this call every node
click. this method can very easily be changed to be called at only the
start of a turn, though, i just got an error with the very first turn of
player 1 and didn't want to have to fix it just yet).