Functional popup menu, and comments ahoy
Created a functional popup menu and began working on the Movement
system. Presently, the popup menu will detect if you have at least one
army on the selected node, and display the movement option if so. If you
click the movement option in the popup menu, it resets the menu to
display all of the armies you control on that node. Presently working on
the system to determine where the selected army can move to.

Also, fully and completely commented out all of my coding on Map.Java.