ToolTip
Added tooltip text to the map image on the map screen. If the mouse is
hovered over the map, it will bring up the text and give the user a
description of what each coloured node represents. Also added the
photoshop file I used to create the buttons, in case we decide to alter
it or add new colours.