Rumours can now be generated
Added time unoccupied property to Nodes in order to capture the effect
of time on nodes with no units occupying it. This is needed for rumours.
Rumour are generated by evaluating the map effects and the properties of
the player, then integrating that information with random variables,
largely normally distributed, in order to generate "false" information.
Rumours become more accurate as time progresses.