More final ui changes for now
okay THESE are the final ui changes for now. It will now inform the
player which type of node was clicked (since i realized the node colour
is the only way of knowing, which if you have an army there it would be
concealed), and changed the wording for actions in the movement log. It
now uses full words within the movement log and within the menus (but
info panel remains letters). Merges and divides will also be listed in
the movement list.