Processes for movement and combat and adjustment complete
The game is now able to process the motions correctly. However, unit
Battle is changed, you will get a Stack Overflow error as the units
participating in combat are not yet removed from the list of events.

Changes to Battle:
call to completeCombat when decision is made
parameters in order:
- ArrayList<CombatUnit> redUnits -> the red faction units in the combat
instance
- Boolean redVictor -> true if the red faction won the battle
- ArrayList<CombatUnit> blueUnits -> the blue faction units in the
combat instance
- Boolean blueVictor -> true if the blue faction won the battle
- Node endLocation -> the node(location) where the winning faction will
terminate their movement
call to successfullFlee when a successful flee is performed
parameters in order:
- ArrayList<CombatUnit> units -> all the units participating in the flee
- Road road -> the Road object upon which the fleeing unit will be
traveling
- int endLocationNum -> the ID of the location that the fleeing unit is
moving to