Full UI Overhaul
Changed the entire UI to make it feel more like a desktop application.
Window borders have returned, all windows are movable but not resizable,
and clicking a map node now makes the popup menu appear wherever the
mouse currently is, as apposed to the button location. All windows
(except the map and map side screen) are now uniformly 600 by 500 in
dimensions and appear in the center of the screen.